code/requirements.txt
gitpython
pygithub
requests
schedule